+++ tests/eval_files/trio101.py
@@ -1,3 +1,4 @@
+# type: ignore
 import contextlib
 import contextlib as bla
 from contextlib import asynccontextmanager, contextmanager, contextmanager as blahabla