+++ tests/eval_files/trio232.py
@@ -1,3 +1,4 @@
+# type: ignore
 import io
 from io import BufferedRandom, BufferedReader, BufferedWriter, TextIOWrapper
 from typing import Any, Optional