+++ tests/eval_files/trio910.py
@@ -1,3 +1,4 @@
+# type: ignore
 import typing
 from typing import Any, overload